# backend/src/api/programs/dto/requests/update-program-request.dto.ts
import { PartialType } from '@nestjs/mapped-types';
import { CreateProgramRequestDto } from './create-program-request.dto';

export class UpdateProgramRequestDto extends PartialType(CreateProgramRequestDto) {}