# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/programs": {
      "get": {
        "operationId": "ProgramsController_getPrograms",
        "parameters": [
          {
            "name": "all",
            "required": false,
            "in": "query",
            "description": "Include all statuses (for admin)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of programs retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all programs",
        "tags": [
          "Programs"
        ]
      },
      "post": {
        "operationId": "ProgramsController_createProgram",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProgramRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Program created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create a new program (admin only)",
        "tags": [
          "Programs"
        ]
      }
    },
    "/api/programs/{id}": {
      "get": {
        "operationId": "ProgramsController_getProgramById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Program ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Program retrieved successfully"
          },
          "404": {
            "description": "Program not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get program by ID",
        "tags": [
          "Programs"
        ]
      },
      "put": {
        "operationId": "ProgramsController_updateProgram",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Program ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgramRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Program updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Program not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Update an existing program (admin only)",
        "tags": [
          "Programs"
        ]
      },
      "delete": {
        "operationId": "ProgramsController_deleteProgram",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Program ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Program deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Program not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Delete a program (admin only - permanent delete)",
        "tags": [
          "Programs"
        ]
      }
    },
    "/api/programs/level/{level}": {
      "get": {
        "operationId": "ProgramsController_getProgramsByLevel",
        "parameters": [
          {
            "name": "level",
            "required": true,
            "in": "path",
            "description": "Program level",
            "schema": {
              "enum": [
                "foundation",
                "diploma",
                "degree"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of programs retrieved successfully"
          },
          "400": {
            "description": "Invalid level"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get programs by level",
        "tags": [
          "Programs"
        ]
      }
    },
    "/api/programs/university/{universityId}": {
      "get": {
        "operationId": "ProgramsController_getProgramsByUniversity",
        "parameters": [
          {
            "name": "universityId",
            "required": true,
            "in": "path",
            "description": "University ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of programs retrieved successfully"
          },
          "400": {
            "description": "Invalid university ID"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get programs by university ID",
        "tags": [
          "Programs"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your Supabase access token"
      }
    },
    "schemas": {
...
      "CreateProgramRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Program name"
          },
          "university_id": {
            "type": "number",
            "description": "University ID"
          },
          "level": {
            "type": "string",
            "description": "Program level",
            "enum": [
              "Foundation",
              "Diploma",
              "Bachelor"
            ]
          },
          "description": {
            "type": "object",
            "description": "Program description"
          },
          "duration_months": {
            "type": "number",
            "description": "Duration in months"
          },
          "tuition_fee_amount": {
            "type": "number",
            "description": "Tuition fee amount"
          },
          "tuition_fee_period": {
            "type": "string",
            "description": "Tuition fee period",
            "enum": [
              "semester",
              "year",
              "total"
            ]
          },
          "currency": {
            "type": "object",
            "description": "Currency code (e.g., MYR)",
            "default": "MYR"
          },
          "start_month": {
            "type": "object",
            "description": "Start month"
          },
          "deadline": {
            "type": "object",
            "description": "Application deadline"
          },
          "field_id": {
            "type": "number",
            "description": "Field of interest ID"
          },
          "entry_requirements": {
            "type": "object",
            "description": "Entry requirements (JSONB)"
          },
          "curriculum": {
            "type": "object",
            "description": "Curriculum (JSONB)"
          },
          "career_outcomes": {
            "type": "object",
            "description": "Career outcomes (JSONB)"
          },
          "facilities": {
            "type": "object",
            "description": "Facilities (JSONB)"
          },
          "tags": {
            "type": "object",
            "description": "Tags (JSONB array)"
          },
          "status": {
            "type": "string",
            "description": "Program status",
            "enum": [
              "active",
              "draft",
              "archived"
            ]
          },
          "rating": {
            "type": "number",
            "description": "Rating"
          },
          "review_count": {
            "type": "number",
            "description": "Review count"
          },
          "employment_rate": {
            "type": "number",
            "description": "Employment rate"
          },
          "average_salary": {
            "type": "number",
            "description": "Average salary"
          },
          "satisfaction_rate": {
            "type": "number",
            "description": "Satisfaction rate"
          },
          "duration": {
            "type": "object",
            "description": "Duration (text format, e.g., \"3 years\")"
          },
          "tuition_fee": {
            "type": "number",
            "description": "Tuition fee (legacy field)"
          },
          "entry_requirements_old": {
            "type": "object",
            "description": "Entry requirements (legacy text field)"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateProgramRequestDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}